+++ src/scss/Home.scss
@@ -1,12 +1,13 @@
 @import './variable.scss';
 .home-container {
   margin-bottom: 10rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 
   .home-text-box {
     text-align: center;
     .desc {
-      color: #999;
-      font-size: 0.87rem;
       margin-bottom: 20px;
 
       p {
@@ -16,10 +17,11 @@
     }
 
     h2 {
-      margin: 15px 0 30px;
+      margin: 15px 0 40px;
       font-size: 1.4rem;
       color: $dark-red;
       font-family: 'HSGaeulSenggak20';
     }
   }
 }
+
