+++ src/scss/button.scss
@@ -1,7 +1,7 @@
 @import './variable.scss';
 
 .button {
-    width: 23rem;
+    width: 20rem;
     display: flex;
     justify-content: center;
     align-items: center;
