+++ src/view/Home.jsx
@@ -1,6 +1,6 @@
 import '../scss/Home.scss';
 import React from 'react';
-import Button from '../comp/Button';
+import Button from '../components/Button';
 import { APP_STATE } from '../constant/stringEnum.js';
 
 const Home = ({ setState }) => {
@@ -14,12 +14,17 @@ const Home = ({ setState }) => {
       <div className="home-container">
         <div className="home-text-box">
           <div className="desc">
-            <p>내 성격과 가장 잘 맞는 강아지 종류는 뭘까?</p>
-            <p>내와 딱 맞는 강아지 찾기 테스트</p>
+            <p className="text--small text--grey">
+              내 성격과 가장 잘 맞는 강아지 종류는 뭘까?
+            </p>
+            <p className="text--small text--grey">
+              내와 딱 맞는 강아지 찾기 테스트
+            </p>
           </div>
           <h2>doggo Finder</h2>
         </div>
         <Button text="START" handler={onStartClick} />
+        <p className="text--smaller text--grey">Designed by ohmjeemin🧡uhjee</p>
       </div>
     </>
   );
